fix: put loading modules in `io/`
